+++ app/src/main/java/com/example/restaurantapp/data/local/dao/DishDao.kt
@@ -13,14 +13,14 @@ interface DishDao {
     suspend fun insertAll(items: List<DishEntity>)
 
     @Query("SELECT * FROM dishes WHERE id = :id LIMIT 1")
-    suspend fun getDishById(id: Int): DishEntity?
+    suspend fun getDishById(id: Long): DishEntity?
 
     @Query("SELECT * FROM dishes WHERE categoryId = :categoryId")
-    fun getDishesByCategoryId(categoryId: Int): Flow<List<DishEntity>>
+    fun getDishesByCategoryId(categoryId: Long): Flow<List<DishEntity>>
 
     @Query("SELECT * FROM dishes WHERE id IN (:ids)")
-    fun observeDishesByIds(ids: List<Int>): Flow<List<DishEntity>>
+    fun observeDishesByIds(ids: List<Long>): Flow<List<DishEntity>>
 
     @Query("SELECT * FROM dishes WHERE restaurantId = :restaurantId")
-    fun observeDishesByRestaurantId(restaurantId: Int): Flow<List<DishEntity>>
+    fun observeDishesByRestaurantId(restaurantId: Long): Flow<List<DishEntity>>
 }
