+++ app/src/main/java/com/example/restaurantapp/data/local/dao/FavoriteSyncDao.kt
@@ -15,10 +15,10 @@ interface FavoriteSyncDao {
     suspend fun getTasks(userId: Long): List<FavoriteSyncEntity>
 
     @Query("SELECT * FROM favorite_sync_queue WHERE userId = :userId AND dishId = :dishId")
-    suspend fun getTask(userId: Long, dishId: Int): FavoriteSyncEntity?
+    suspend fun getTask(userId: Long, dishId: Long): FavoriteSyncEntity?
 
     @Query("DELETE FROM favorite_sync_queue WHERE userId = :userId AND dishId = :dishId")
-    suspend fun removeTask(userId: Long, dishId: Int)
+    suspend fun removeTask(userId: Long, dishId: Long)
 
     @Query("DELETE FROM favorite_sync_queue WHERE userId = :userId")
     suspend fun clearTasks(userId: Long)
