+++ app/src/main/java/com/example/restaurantapp/data/remote/dto/request/ReservationRequest.kt
@@ -1,7 +1,7 @@
 package com.example.restaurantapp.data.remote.dto.request
 
 data class ReservationRequest(
-    val restaurantId: Int,
+    val restaurantId: Long,
     val dateTime: String,
     val guests: Int
 )
