+++ app/src/main/java/com/example/restaurantapp/data/remote/dto/response/NewsResponse.kt
@@ -2,7 +2,7 @@ package com.example.restaurantapp.data.remote.dto.response
 
 data class NewsResponse(
     val id: Long,
-    val restaurantId: Int?,
+    val restaurantId: Long?,
     val title: String,
     val content: String,
     val createdAt: String
