+++ app/src/main/java/com/example/restaurantapp/data/remote/dto/response/ReservationResponse.kt
@@ -2,7 +2,7 @@ package com.example.restaurantapp.data.remote.dto.response
 
 data class ReservationResponse(
     val id: Long,
-    val restaurantId: Int,
+    val restaurantId: Long,
     val tableId: Long,
     val startTime: String,
     val endTime: String,
