+++ app/src/main/java/com/example/restaurantapp/domain/repository/CategoriesRepository.kt
@@ -4,6 +4,6 @@ import com.example.restaurantapp.domain.model.Category
 import kotlinx.coroutines.flow.Flow
 
 interface CategoriesRepository {
-    fun getCategoriesByRestaurantId(restaurantId: Int): Flow<List<Category>>
-    suspend fun getCategoryById(id: Int): Category?
+    fun getCategoriesByRestaurantId(restaurantId: Long): Flow<List<Category>>
+    suspend fun getCategoryById(id: Long): Category?
 }
