+++ app/src/main/java/com/example/restaurantapp/domain/repository/DishesRepository.kt
@@ -4,8 +4,8 @@ import com.example.restaurantapp.domain.model.Dish
 import kotlinx.coroutines.flow.Flow
 
 interface DishesRepository {
-    fun getDishesByCategoryId(categoryId: Int): Flow<List<Dish>>
-    fun observeDishesByIds(ids: List<Int>): Flow<List<Dish>>
-    fun observeDishesByRestaurantId(restaurantId: Int): Flow<List<Dish>>
-    suspend fun getDishById(id: Int): Dish?
+    fun getDishesByCategoryId(categoryId: Long): Flow<List<Dish>>
+    fun observeDishesByIds(ids: List<Long>): Flow<List<Dish>>
+    fun observeDishesByRestaurantId(restaurantId: Long): Flow<List<Dish>>
+    suspend fun getDishById(id: Long): Dish?
 }
