+++ app/src/main/java/com/example/restaurantapp/domain/repository/FavoriteDishRepository.kt
@@ -4,10 +4,10 @@ import kotlinx.coroutines.flow.Flow
 
 interface FavoriteDishRepository {
     fun isAuthorized(): Boolean
-    fun observeIsFavorite(dishId: Int): Flow<Boolean>
-    fun observeFavoriteDishIds(): Flow<List<Int>>
-    suspend fun addToFavorites(dishId: Int): Result<Unit>
-    suspend fun removeFromFavorites(dishId: Int): Result<Unit>
+    fun observeIsFavorite(dishId: Long): Flow<Boolean>
+    fun observeFavoriteDishIds(): Flow<List<Long>>
+    suspend fun addToFavorites(dishId: Long): Result<Unit>
+    suspend fun removeFromFavorites(dishId: Long): Result<Unit>
     suspend fun pushFavorites(): Result<Unit>
     suspend fun pullFavorites(): Result<Unit>
     suspend fun syncFavorites(): Result<Unit>
