+++ app/src/main/java/com/example/restaurantapp/domain/repository/NewsRepository.kt
@@ -3,5 +3,5 @@ package com.example.restaurantapp.domain.repository
 import com.example.restaurantapp.domain.model.News
 
 interface NewsRepository {
-    suspend fun getNews(restaurantId: Int? = null): Result<List<News>>
+    suspend fun getNews(restaurantId: Long? = null): Result<List<News>>
 }
