+++ app/src/main/java/com/example/restaurantapp/presentation/menu/MenuCategoriesFragment.kt
@@ -51,7 +51,7 @@ class MenuCategoriesFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val restaurantId = arguments?.getInt("restaurantId") ?: return
+        val restaurantId = arguments?.getLong("restaurantId") ?: return
         
         setupRecyclerViews()
         setupSearchView()
