+++ app/src/main/java/com/example/restaurantapp/presentation/reservation/ReservationFragment.kt
@@ -46,7 +46,7 @@ class ReservationFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val restaurantId = arguments?.getInt("restaurantId") ?: return
+        val restaurantId = arguments?.getLong("restaurantId") ?: return
         viewModel.loadRestaurant(restaurantId)
 
         setupRecyclerViews()
